--- src/main/resources/templates/admin/phieu_giam/cart.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>Sản phẩm giảm giá</title>
-</head>
-<body>
-<h1>Các sản phẩm áp dụng phiếu giảm giá</h1>
-<table border="1">
-    <thead>
-    <tr>
-        <th>Tên sản phẩm</th>
-        <th>Số tiền giảm</th>
-        <th>Giá sản phẩm</th>
-        <th>Số tiền sau giảm</th>
-    </tr>
-    </thead>
-    <tbody>
-    <tr th:each="item : ${discountProducts}">
-        <td th:text="${item.sanPham.ten}">Tên sản phẩm</td>
-        <td th:text="${item.PhieuGiam.giaTriGiam}">Giá trị giảm</td>
-        <td th:text="${item.sanPhamChiTiet.gia}">Giá sản phẩm</td>
-        <td th:text="${item.sanPhamChiTiet.gia - item.PhieuGiam.giaTriGiam}"> <!-- Tính toán giá sau giảm -->
-            Giá sau giảm
-        </td>
-    </tr>
-    </tbody>
-</table>
-</body>
-</html>
