--- src/main/resources/templates/admin/phieu_giam/test.html
@@ -1,95 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head th:replace="/admin/fragments/head :: head" xmlns:th="http://www.thymeleaf.org">
-    <link rel="stylesheet" href="assets/extensions/toastify-js/src/toastify.css">
-</head>
-<body>
-<!-- Site wrapper -->
-<div class="wrapper">
-    <!-- Sidebar -->
-    <div>
-        <aside th:replace="/admin/fragments/sidebar :: sidebar" ></aside>
-    </div>
-
-    <!-- Main-->
-    <div id="main">
-        <header th:replace="/admin/fragments/header :: header"></header>
-        <!--        <header class="mb-3">-->
-        <!--            <a href="#" class="burger-btn d-block d-xl-none">-->
-        <!--                <i class="bi bi-justify fs-3"></i>-->
-        <!--            </a>-->
-        <!--        </header>-->
-
-        <div class="page-heading">
-            <div class="page-title">
-                <div class="row">
-                    <div class="col-12 col-md-6 order-md-1 order-last">
-                        <h3>Khung, thay tiêu đề ở đây</h3>
-                        <p class="text-subtitle text-muted">chữ nhỏ, k dùng thì xóa</p>
-                    </div>
-                    <div class="col-12 col-md-6 order-md-2 order-first">
-                        <nav aria-label="breadcrumb" class="breadcrumb-header float-start float-lg-end">
-                            <ol class="breadcrumb">
-                                <li class="breadcrumb-item"><a href="/admin/index">Home</a></li>
-                                <li class="breadcrumb-item active" aria-current="page"><a href="#">Quản lý ...</a></li>
-                            </ol>
-                        </nav>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <div class="page-content">
-            <!--        dữ liệu thực hiện trong secsion-->
-            <section class="section">
-                <div class="row">
-                    <div class="col-12">
-                        <div class="card">
-                            <div class="card-header">
-                                <h4 class="card-title">Position</h4>
-                            </div>
-                            <div class="card-body">
-                                <div class="row">
-                                    <div class="col-12 col-md-4">
-                                        <button id="top-left" class="btn btn-outline-primary btn-block btn-lg">Top Left</button>
-                                    </div>
-                                    <div class="col-12 col-md-4">
-                                        <button id="top-center" class="btn btn-outline-primary btn-block btn-lg">Top
-                                            Center</button>
-                                    </div>
-                                    <div class="col-12 col-md-4">
-                                        <button id="top-right" class="btn btn-outline-primary btn-block btn-lg">Top
-                                            Right</button>
-                                    </div>
-                                </div>
-                                <div class="row mt-3">
-                                    <div class="col-12 col-md-4">
-                                        <button id="bottom-left" class="btn btn-outline-primary btn-block btn-lg">Bottom
-                                            Left</button>
-                                    </div>
-                                    <div class="col-12 col-md-4">
-                                        <button id="bottom-center" class="btn btn-outline-primary btn-block btn-lg">Bottom
-                                            Center</button>
-                                    </div>
-                                    <div class="col-12 col-md-4">
-                                        <button id="bottom-right" class="btn btn-outline-primary btn-block btn-lg">Bottom
-                                            Right</button>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </section>
-        </div>
-    </div>
-    <!-- Footer-->
-    <div>
-        <footer th:replace="/admin/fragments/footer :: footer"></footer>
-    </div>
-</div>
-<div>
-    <script th:replace="/admin/fragments/script :: script"></script>
-</div>
-</body>
-</html>
