+++ src/main/resources/templates/admin/website/orderInfor.html
@@ -307,7 +307,7 @@
     <div class="order-summary">
         <h3>Giỏ hàng của bạn</h3>
         <div th:each="item : ${cartItems}" class="product-item">
-            <img src="https://via.placeholder.com/60" alt="Product Image">
+            <img src="https://gamek.mediacdn.vn/133514250583805952/2021/10/3/ani3-16332425354691599897585.jpg" alt="Product Image">
             <span class="badge-quantity" th:text="${item.quantity}">1</span> <!-- Quantity Badge -->
             <div class="product-info">
                 <p th:text="${item.name}">Tên sản phẩm</p>
