--- src/main/resources/templates/admin/website/product-catalog.css
@@ -1,172 +1,0 @@
-@charset "UTF-8";
-
-body {
-    font-family: Arial, sans-serif;
-}
-
-.navbar-custom {
-    background-color: #ffffff; /* Màu nền */
-    border-bottom: 1px solid #ddd; /* Đường viền dưới */
-}
-
-.navbar-brand img {
-    transition: transform 0.3s; /* Hiệu ứng zoom khi di chuột */
-}
-
-.navbar-brand img:hover {
-    transform: scale(1.1); /* Phóng to logo khi di chuột */
-}
-
-.search-bar {
-    position: relative;
-    width: 350px; /* Độ rộng của thanh tìm kiếm */
-}
-
-.search-bar input {
-    width: 100%;
-    padding: 10px; /* Đệm trong */
-    border: 1px solid #ccc; /* Đường viền */
-    border-radius: 20px; /* Bo tròn góc */
-    outline: none; /* Bỏ viền khi được chọn */
-    transition: border-color 0.3s; /* Hiệu ứng chuyển đổi màu viền */
-}
-
-.search-bar input:focus {
-    border-color: #007bff; /* Màu viền khi được chọn */
-}
-
-.search-bar button {
-    position: absolute;
-    right: 0;
-    top: 50%;
-    transform: translateY(-50%);
-    background-color: #007bff; /* Màu nền nút */
-    color: white; /* Màu chữ */
-    border: none; /* Bỏ đường viền */
-    border-radius: 20px; /* Bo tròn góc */
-    padding: 10px 15px; /* Đệm trong */
-    cursor: pointer; /* Con trỏ khi hover */
-    transition: background-color 0.3s; /* Hiệu ứng chuyển đổi màu nền */
-}
-
-.search-bar button:hover {
-    background-color: #0056b3; /* Màu nền khi hover */
-}
-
-.contact-info {
-    color: #333; /* Màu chữ */
-    font-size: 14px; /* Kích thước chữ */
-}
-
-.icons-section a {
-    color: #333; /* Màu chữ cho biểu tượng */
-    font-size: 20px; /* Kích thước biểu tượng */
-    margin-left: 15px; /* Khoảng cách giữa các biểu tượng */
-    transition: color 0.3s; /* Hiệu ứng chuyển đổi màu */
-}
-
-.icons-section a:hover {
-    color: #007bff; /* Màu chữ khi hover */
-}
-
-.product-card {
-    transition: transform 0.3s ease-in-out;
-    position: relative;
-    overflow: hidden; /* Đảm bảo nội dung không ra ngoài card */
-}
-
-.product-card:hover {
-    transform: scale(1.05);
-    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.2); /* Hiệu ứng bóng */
-}
-
-.product-title {
-    min-height: 50px;
-    font-weight: bold; /* Đậm tiêu đề sản phẩm */
-}
-
-.product-price {
-    font-size: 1.25rem;
-    font-weight: bold;
-}
-
-.old-price {
-    text-decoration: line-through;
-    color: #888;
-    margin-right: 10px; /* Khoảng cách với giá mới */
-}
-
-.discount-badge {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    background-color: red;
-    color: white;
-    padding: 5px;
-    font-size: 0.9rem;
-    border-radius: 5px;
-}
-
-footer {
-    background-color: #343a40;
-    color: #fff;
-}
-
-.carousel-item img {
-    width: 100%;
-    height: auto;
-}
-
-#carouselExample {
-    width: 80%; /* Chiều rộng 80% để slideshow hẹp lại */
-    max-width: 100%; /* Đảm bảo không vượt quá 100% */
-    height: 300px; /* Chiều cao bạn mong muốn, có thể điều chỉnh thêm */
-    overflow: hidden; /* Ẩn phần thừa */
-    margin: auto; /* Căn giữa slideshow */
-}
-
-#carouselExample .carousel-item {
-    height: 100%; /* Đảm bảo carousel-item có chiều cao đầy đủ */
-}
-
-#carouselExample .carousel-item img {
-    height: 100%; /* Đảm bảo ảnh đầy chiều cao của carousel */
-    object-fit: cover; /* Giữ tỷ lệ ảnh, có thể cắt bớt */
-    width: 100%; /* Đảm bảo ảnh chiếm toàn bộ chiều rộng */
-}
-
-.footer-info {
-    font-size: 14px; /* Thay đổi kích thước chữ nếu cần */
-    color: #6c757d; /* Màu chữ nếu muốn */
-}
-
-.dropdown-menu {
-    display: none; /* Ẩn menu con mặc định */
-}
-
-.dropdown:hover .dropdown-menu {
-    display: block; /* Hiển thị menu con khi di chuột */
-}
-.dropdown-item img {
-    width: 30px; /* Điều chỉnh kích thước hình ảnh */
-    margin-right: 10px; /* Khoảng cách giữa hình ảnh và văn bản */
-}
-#navbar2 {
-    background-color: #007BFF; /* Màu nền */
-}
-
-#navbar2 .navbar-nav .nav-link {
-    color: #ffffff; /* Màu chữ trắng */
-}
-
-#navbar2 .navbar-nav .dropdown-menu {
-    background-color: #0056b3; /* Màu nền cho menu dropdown */
-}
-
-#navbar2 .navbar-nav .dropdown-item {
-    color: #ffffff; /* Màu chữ cho items trong dropdown */
-}
-
-#navbar2 .navbar-nav .dropdown-item:hover {
-    background-color: #004085; /* Màu nền khi hover */
-}
